WIP Refactor UI
