+++ src/main/java/de/embl/cba/mobie/image/ImagePropertiesEditor.java
@@ -1,7 +1,7 @@
 package de.embl.cba.mobie.image;
 
 import de.embl.cba.mobie.projects.ProjectsCreator;
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.ProjectManager;
 import de.embl.cba.mobie.utils.Utils;
 import de.embl.cba.tables.FileAndUrlUtils;
 import de.embl.cba.tables.SwingUtils;
@@ -19,7 +19,6 @@ import java.text.Format;
 import java.text.NumberFormat;
 import java.text.ParseException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -75,7 +74,7 @@ public class ImagePropertiesEditor {
     private JPanel createComboPanel (JComboBox<String> combo, String label) {
         final JPanel comboPanel = SwingUtils.horizontalLayoutPanel();
         setComboBoxDimensions( combo );
-        combo.setPrototypeDisplayValue(MoBIEViewer.PROTOTYPE_DISPLAY_VALUE);
+        combo.setPrototypeDisplayValue( ProjectManager.PROTOTYPE_DISPLAY_VALUE);
         comboPanel.add(getJLabel(label));
         comboPanel.add(combo);
