+++ src/main/java/de/embl/cba/mobie/ui/command/OpenPlatyBrowserCommand.java
@@ -1,6 +1,6 @@
 package de.embl.cba.mobie.ui.command;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.ProjectManager;
 import org.scijava.command.Command;
 import org.scijava.plugin.Plugin;
 
@@ -13,6 +13,6 @@ public class OpenPlatyBrowserCommand implements Command
 	@Override
 	public void run()
 	{
-		new MoBIEViewer( githubProject );
+		new ProjectManager( githubProject );
 	}
 }
