+++ src/test/java/develop/DevelopAnnotationCorrectionPlatynereis.java
@@ -1,12 +1,12 @@
 package develop;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.ProjectManager;
 
 public class DevelopAnnotationCorrectionPlatynereis
 {
 	public static void main( String[] args )
 	{
-		final MoBIEViewer viewer = new MoBIEViewer( "https://github.com/vzinche/platybrowser-backend" );
+		final ProjectManager viewer = new ProjectManager( "https://github.com/vzinche/platybrowser-backend" );
 		viewer.getSourcesManager().addSourceToPanelAndViewer( "sbem-6dpf-1-whole-segmented-cells" );
 
 		// TODO: develop API get the views of a source: getViews( String sourceName )
