+++ src/test/java/projects/OpenLocalCovidEMCF.java
@@ -1,11 +1,11 @@
 package projects;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.ProjectManager;
 
 public class OpenLocalCovidEMCF
 {
 	public static void main( String[] args )
 	{
-		new MoBIEViewer( "/Volumes/emcf/common/5792_Sars-Cov-2/covid-em/data");
+		new ProjectManager( "/Volumes/emcf/common/5792_Sars-Cov-2/covid-em/data");
 	}
 }
