+++ src/test/java/projects/OpenLocalPlankton.java
@@ -1,7 +1,7 @@
 package projects;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEOptions;
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.MoBIEOptions;
+import de.embl.cba.mobie.ui.ProjectManager;
 import net.imagej.ImageJ;
 
 public class OpenLocalPlankton
@@ -9,6 +9,6 @@ public class OpenLocalPlankton
 	public static void main( String[] args )
 	{
 		new ImageJ().ui().showUI();
-		new MoBIEViewer("/Volumes/schwab/schwab/MoBIE-plankton", MoBIEOptions.options().imageDataStorageModality( MoBIEOptions.ImageDataStorageModality.FileSystem ));
+		new ProjectManager("/Volumes/schwab/schwab/MoBIE-plankton", MoBIEOptions.options().imageDataStorageModality( MoBIEOptions.ImageDataStorageModality.FileSystem ));
 	}
 }
