+++ src/test/java/projects/OpenLocalPlatynereis.java
@@ -1,11 +1,11 @@
 package projects;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
+import de.embl.cba.mobie.ui.ProjectManager;
 
 public class OpenLocalPlatynereis
 {
 	public static void main( String[] args )
 	{
-		new MoBIEViewer("/g/arendt/EM_6dpf_segmentation/platy-browser-data/data/" );
+		new ProjectManager("/g/arendt/EM_6dpf_segmentation/platy-browser-data/data/" );
 	}
 }
