+++ src/test/java/projects/OpenRemotePlatynereisBranch.java
@@ -1,12 +1,11 @@
 package projects;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
-import de.embl.cba.mobie.ui.viewer.MoBIEOptions;
+import de.embl.cba.mobie.ui.ProjectManager;
 
 public class OpenRemotePlatynereisBranch
 {
 	public static void main( String[] args )
 	{
-		new MoBIEViewer( "https://github.com/platybrowser/platybrowser" );
+		new ProjectManager( "https://github.com/platybrowser/platybrowser" );
 	}
 }
