+++ src/test/java/publication/platybrowser/RegistrationVisualisationMuscles.java
@@ -1,7 +1,8 @@
 package publication.platybrowser;
 
-import de.embl.cba.mobie.ui.viewer.MoBIEViewer;
-import de.embl.cba.mobie.ui.viewer.SourcesManager;
+import de.embl.cba.mobie.ui.ProjectManager;
+import de.embl.cba.mobie.ui.SourcesManager;
+import de.embl.cba.mobie.utils.UniverseUtils;
 import net.imagej.ImageJ;
 
 import java.awt.*;
@@ -12,11 +13,11 @@ public class RegistrationVisualisationMuscles
 	{
 		new ImageJ().ui().showUI();
 
-		final MoBIEViewer moBIEViewer = new MoBIEViewer(
+		final ProjectManager projectManager = new ProjectManager(
 				"/Volumes/arendt/EM_6dpf_segmentation/platy-browser-data/data",
 				"https://git.embl.de/tischer/platy-browser-tables/raw/dev/data" );
 
-		final SourcesManager sourcesManager = moBIEViewer.getSourcesManager();
+		final SourcesManager sourcesManager = projectManager.getSourcesManager();
 
 		sourcesManager.addSourceToPanelAndViewer( "sbem-6dpf-1-whole-segmented-muscles" );
 		sourcesManager.addSourceToPanelAndViewer( "prospr-6dpf-1-whole-MHCL4-MED" );
@@ -29,14 +30,14 @@ public class RegistrationVisualisationMuscles
 
 		// OVERVIEW
 
-		moBIEViewer.getUserInterfaceProvider().setView( "View: (-0.747871017528087, -0.6676436522042726, 2.0869538546605892, 291.53347494758805, 1.4642245659643167, -1.7927905904117456, -0.04882433847545138, 337.95348110873715, 1.6300830832316395, 1.3040664665853112, 1.0013367511280067, -535.9999606550252)" );
+		projectManager.getUserInterfacePanelsProvider().setView( "View: (-0.747871017528087, -0.6676436522042726, 2.0869538546605892, 291.53347494758805, 1.4642245659643167, -1.7927905904117456, -0.04882433847545138, 337.95348110873715, 1.6300830832316395, 1.3040664665853112, 1.0013367511280067, -535.9999606550252)" );
 
 		// TODO: get this via keyboard shortcut
-		moBIEViewer.getUserInterfaceProvider().setVolumeView("center =  1.0 0.0 0.0 122.88 0.0 1.0 0.0 122.88 0.0 0.0 1.0 140.40001 0.0 0.0 0.0 1.0\n" +
+		UniverseUtils.setVolumeView("center =  1.0 0.0 0.0 122.88 0.0 1.0 0.0 122.88 0.0 0.0 1.0 140.40001 0.0 0.0 0.0 1.0\n" +
 				"translate =  1.0 0.0 0.0 0.117842264 0.0 1.0 0.0 0.15685147 0.0 0.0 1.0 -0.52159584 0.0 0.0 0.0 1.0\n" +
 				"rotate =  -0.22049984 -0.6480566 -0.72897357 0.0 -0.22329058 0.76105946 -0.60904 0.0 0.94948465 0.028479699 -0.31251845 0.0 0.0 0.0 0.0 1.0\n" +
 				"zoom =  1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 1.0 358.10345 0.0 0.0 0.0 1.0\n" +
-				"animate =  1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 1.0\n"
+				"animate =  1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 1.0\n", projectManager.getUserInterfacePanelsProvider().sourcesManager.getUniverse()
 		);
